+++ pkg/querier/engine/query_tracker.go
@@ -6,9 +6,8 @@ import (
 	"bytes"
 	"context"
 
-	"github.com/weaveworks/common/tracing"
-
 	"github.com/grafana/dskit/tenant" //lint:ignore faillint queryTracker needs tenant package
+	"github.com/weaveworks/common/tracing"
 
 	"github.com/grafana/mimir/pkg/util/activitytracker" //lint:ignore faillint queryTracker needs activitytracker
 )