+++ pkg/util/limiter/utilization.go
@@ -11,6 +11,8 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/services"
 	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/procfs"
 	"go.uber.org/atomic"
 
@@ -64,10 +66,12 @@ type UtilizationBasedLimiter struct {
 	lastUpdate     time.Time
 	cpuMovingAvg   *math.EwmaRate
 	limitingReason atomic.String
+	currCPUUtil    atomic.Float64
+	currMemoryUtil atomic.Uint64
 }
 
 // NewUtilizationBasedLimiter returns a UtilizationBasedLimiter configured with cpuLimit and memoryLimit.
-func NewUtilizationBasedLimiter(cpuLimit float64, memoryLimit uint64, logger log.Logger) *UtilizationBasedLimiter {
+func NewUtilizationBasedLimiter(cpuLimit float64, memoryLimit uint64, logger log.Logger, reg prometheus.Registerer) *UtilizationBasedLimiter {
 	// Calculate alpha for a minute long window
 	// https://github.com/VividCortex/ewma#choosing-alpha
 	alpha := 2 / (resourceUtilizationSlidingWindow.Seconds()/resourceUtilizationUpdateInterval.Seconds() + 1)
@@ -79,6 +83,22 @@ func NewUtilizationBasedLimiter(cpuLimit float64, memoryLimit uint64, logger log
 		cpuMovingAvg: math.NewEWMARate(alpha, resourceUtilizationUpdateInterval),
 	}
 	l.Service = services.NewTimerService(resourceUtilizationUpdateInterval, l.starting, l.update, nil)
+
+	if reg != nil {
+		promauto.With(reg).NewGaugeFunc(prometheus.GaugeOpts{
+			Name: "utilization_limiter_current_cpu_load",
+			Help: "Current average CPU load calculated by utilization based limiter.",
+		}, func() float64 {
+			return l.currCPUUtil.Load()
+		})
+		promauto.With(reg).NewGaugeFunc(prometheus.GaugeOpts{
+			Name: "utilization_limiter_current_memory_usage_bytes",
+			Help: "Current memory usage calculated by utilization based limiter.",
+		}, func() float64 {
+			return float64(l.currMemoryUtil.Load())
+		})
+	}
+
 	return l
 }
 
@@ -101,13 +121,13 @@ func (l *UtilizationBasedLimiter) starting(_ context.Context) error {
 }
 
 func (l *UtilizationBasedLimiter) update(_ context.Context) error {
-	l.compute(time.Now())
+	l.compute(time.Now)
 	return nil
 }
 
 // compute and return the current CPU and memory utilization.
 // This function must be called at a regular interval (resourceUtilizationUpdateInterval) to get a predictable behaviour.
-func (l *UtilizationBasedLimiter) compute(now time.Time) (currCPUUtil float64, currMemoryUtil uint64) {
+func (l *UtilizationBasedLimiter) compute(nowFn func() time.Time) (currCPUUtil float64, currMemoryUtil uint64) {
 	cpuTime, currMemoryUtil, err := l.utilizationScanner.Scan()
 	if err != nil {
 		level.Warn(l.logger).Log("msg", "failed to get CPU and memory stats", "err", err.Error())
@@ -115,6 +135,11 @@ func (l *UtilizationBasedLimiter) compute(now time.Time) (currCPUUtil float64, c
 		l.limitingReason.Store("")
 		return
 	}
+
+	// Get wall time after CPU time, in case there's a delay before CPU time is returned,
+	// which would cause us to compute too high of a CPU load
+	now := nowFn()
+	l.currMemoryUtil.Store(currMemoryUtil)
 
 	// Add the instant CPU utilization to the moving average. The instant CPU
 	// utilization can only be computed starting from the 2nd tick.
@@ -133,7 +158,8 @@ func (l *UtilizationBasedLimiter) compute(now time.Time) (currCPUUtil float64, c
 	if l.firstUpdate.IsZero() {
 		l.firstUpdate = now
 	} else if now.Sub(l.firstUpdate) >= resourceUtilizationSlidingWindow {
-		currCPUUtil = float64(l.cpuMovingAvg.Rate()) / 100
+		currCPUUtil = l.cpuMovingAvg.Rate() / 100
+		l.currCPUUtil.Store(currCPUUtil)
 	}
 
 	var reason string